move all the HDF5 particle routines to a separate file
