+++ Src/Particle/CMakeLists.txt
@@ -38,6 +38,7 @@ target_sources( amrex
    AMReX_ParticleMesh.H
    AMReX_ParticleLocator.H
    AMReX_ParticleIO.H
+   AMReX_ParticleHDF5.H
    AMReX_DenseBins.H
    AMReX_BinIterator.H
    AMReX_ParticleTransformation.H
